+++ packages/components/src/dropdown-menu-v2/types.ts
@@ -249,7 +249,7 @@ export type DropdownMenuGroupProps = {
 
 export type DropdownMenuSeparatorProps = {};
 
-export type DropdownMenuContext = {
+export type DropdownMenuInternalContext = {
 	/**
 	 * This variant can be used to change the appearance of the component in
 	 * specific contexts, ie. when rendered inside the `Toolbar` component.
@@ -258,7 +258,7 @@ export type DropdownMenuContext = {
 };
 
 export type DropdownMenuPrivateContext = Pick<
-	DropdownMenuContext,
+	DropdownMenuInternalContext,
 	'variant'
 > & {
 	portalContainer: HTMLElement | null;